+++ src/Microsoft.Performance.SDK.Tests/FileDataSourceAttributeTests.cs
@@ -108,55 +108,4 @@ namespace Microsoft.Performance.SDK.Tests
             Assert.IsTrue(sut.Accepts(dataSource));
         }
     }
-
-    [TestClass]
-    public class ExtensionlessFileDataSourceAttributeTests
-    {
-        [TestMethod]
-        [UnitTest]
-        public void NullDescriptionThrows()
-        {
-            Assert.ThrowsException<ArgumentNullException>(
-                () => new ExtensionlessFileDataSourceAttribute(null));
-        }
-
-        [TestMethod]
-        [UnitTest]
-        public void WhitespaceDescriptionThrows()
-        {
-            Assert.ThrowsException<ArgumentException>(
-                () => new ExtensionlessFileDataSourceAttribute(string.Empty));
-        }
-
-        [TestMethod]
-        [UnitTest]
-        public void DescriptionIsSetWhenProvided()
-        {
-            var description = "This is a test.";
-
-            var sut = new ExtensionlessFileDataSourceAttribute(description);
-
-            Assert.AreEqual(description, sut.Description);
-        }
-
-        [TestMethod]
-        [UnitTest]
-        public void AcceptsFailsIfHasExtension()
-        {
-            var dataSource = new FileDataSource(@"c:\test\bad.ext");
-            var sut = new ExtensionlessFileDataSourceAttribute("Empty");
-
-            Assert.IsFalse(sut.Accepts(dataSource));
-        }
-
-        [TestMethod]
-        [UnitTest]
-        public void AcceptsSucceedsIfDoesNotHaveExtension()
-        {
-            var dataSource = new FileDataSource(@"c:\test\good");
-            var sut = new ExtensionlessFileDataSourceAttribute("Empty");
-
-            Assert.IsTrue(sut.Accepts(dataSource));
-        }
-    }
 }
